Extract the `/verticals` endpoint interacting functions out
