+++ client/state/signup/steps/site-vertical/constants.js
@@ -1,0 +1,2 @@
+export const SITE_VERTICALS_REQUEST_ID = 'site-verticals-search-results';
+export const DEFAULT_SITE_VERTICAL_REQUEST_ID = 'default-site-verticals-search-results';
