+++ client/state/signup/steps/site-vertical/selectors.js
@@ -6,6 +6,12 @@
 
 import { get } from 'lodash';
 
+/**
+ * Internal dependencies
+ */
+import { getHttpData } from 'state/data-layer/http-data';
+import { SITE_VERTICALS_REQUEST_ID, DEFAULT_SITE_VERTICAL_REQUEST_ID } from './constants';
+
 export function getSiteVerticalName( state ) {
 	return get( state, 'signup.steps.siteVertical.name', '' );
 }
@@ -29,3 +35,23 @@ export function getSiteVerticalIsUserInput( state ) {
 export function getSiteVerticalPreview( state ) {
 	return get( state, 'signup.steps.siteVertical.preview', '' );
 }
+
+export const isVerticalSearchPending = () =>
+	'pending' === get( getHttpData( SITE_VERTICALS_REQUEST_ID ), 'state', false );
+
+export const getSiteVerticalsLastUpdated = () => {
+	const siteVerticalHttpData = getHttpData( SITE_VERTICALS_REQUEST_ID );
+
+	return get( siteVerticalHttpData, 'lastUpdated', 0 );
+};
+
+export const getSiteVerticals = () => {
+	const siteVerticalHttpData = getHttpData( SITE_VERTICALS_REQUEST_ID );
+
+	return get( siteVerticalHttpData, 'data', [] );
+};
+
+export const getDefaultVertical = () => {
+	const defaultVerticalHttpData = getHttpData( DEFAULT_SITE_VERTICAL_REQUEST_ID );
+	return get( defaultVerticalHttpData, 'data[0]', {} );
+};
